+++ lib/pgsync.rb
@@ -93,8 +93,10 @@ module PgSync
         if opts[:schema_only]
           log "* Dumping schema"
           tables = tables.keys.map { |t| "-t #{t}" }.join(" ")
-          dump_command = "pg_dump --verbose --schema-only --no-owner --no-acl --clean #{tables} #{to_url(source_uri)}"
-          restore_command = "psql -q -d #{to_url(destination_uri)}"
+          psql_version = Gem::Version.new(`psql --version`.split(" ").last)
+          if_exists = psql_version >= Gem::Version.new("9.4.0")
+          dump_command = "pg_dump -Fc --verbose --schema-only --no-owner --no-acl #{tables} #{to_url(source_uri)}"
+          restore_command = "pg_restore --no-owner --no-acl --clean #{if_exists ? "--if-exists" : nil} -d #{to_url(destination_uri)}"
           system("#{dump_command} | #{restore_command}")
 
           log_completed(start_time)